structurers: Add UdataStructurer; tests: Fix random fails
Adds the UdataStructurer to the structurers module. This also fixes
an issue with the tests, where if additional checks were performed
(.e.g. in case array ordering is different between the expected and
the actual), these would fail on items with a 'none' value, because
the 'assert_none' (& 'all_fields') parameter was not passed to
nested function calls