Fix line locations being considered valid
These are just straight lines, that are erroneously added as a
bounding box. Previously they were stored as envelope if both x
values or both y values were the same, but not x and y at the same
time. With the new code, this is not valid anymore, and does not
get stored