Add GPLv3 License statements
Optimally it would use a license like LGPL or OSL, so the software
without changes, can still be linked from other programs. However,
because the 'unidecode' python package is licensed under GPLv2
or later, and 'html2text' is licensed under LGPLv3, this last
license is the only resort. AGPL can also not be used because of
imcompatibility with GPLv2